Moving files from the Old Repository

+++ src/w_icons.c
@@ -3241,9 +3241,9 @@ char * fig_reduced_c_icon_X[] = {
 
 #include "splash.xbm"
 
-icon_struct	letters_ic = { (int) letters_width, (int) letters_height,
+icon_struct	letters_ic = { letters_width, letters_height,
 			      (char*)letters_bits };
-icon_struct	spl_bckgnd_ic = { (int) spl_bckgnd_width, (int) spl_bckgnd_height,
+icon_struct	spl_bckgnd_ic = { spl_bckgnd_width, spl_bckgnd_height,
 			      (char*)spl_bckgnd_bits };
 
 #define		MODE_SW_HT_SMALL	22	/* height of a small mode switch icon */
